+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## v1.2.0 (TBD)
 
+- Every command line option now has a short form.
 - Errors from voices are no longer printed unless at no voice yields any audio.
 - Updated documentation.